refactor(benchmarks): own raster conformance captures
