+++ apps/benchmarks/src/benchmark/targets/conformance/raster/contracts.ts
@@ -16,8 +16,8 @@ export interface RasterSourceOutlineCapture {
 }
 
 /**
- * Backend-private resources stay in their renderer module. The target owns this
- * session's warm load, capture, and disposal lifecycle without owning a renderer.
+ * Technique-private resources stay in the selected target module. The target owns
+ * this session's warm load, capture, and disposal lifecycle without owning a renderer.
  */
 export interface RasterConformanceSession {
   load(input: BenchmarkInput, controls: BenchmarkControls, context?: BenchmarkExecutionContext): Promise<void>;
